requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
appnope==0.1.0            # via ipykernel, ipython
attrs==19.3.0             # via jsonschema
backcall==0.2.0           # via ipython
beautifulsoup4==4.9.1     # via -r requirements.in
bleach==3.1.5             # via nbconvert
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
cycler==0.10.0            # via matplotlib
decorator==4.4.2          # via ipython, traitlets
defusedxml==0.6.0         # via nbconvert
entrypoints==0.3          # via nbconvert
idna==2.10                # via requests
importlib-metadata==1.7.0  # via jsonschema
ipykernel==5.3.3          # via notebook
ipython-genutils==0.2.0   # via nbformat, notebook, traitlets
ipython==7.16.1           # via ipykernel
jedi==0.17.2              # via ipython
jinja2==2.11.2            # via jupyterlab, jupyterlab-server, nbconvert, notebook
json5==0.9.5              # via jupyterlab-server
jsonschema==3.2.0         # via jupyterlab-server, nbformat
jupyter-client==6.1.6     # via ipykernel, notebook
jupyter-core==4.6.3       # via jupyter-client, nbconvert, nbformat, notebook
jupyterlab-server==1.2.0  # via jupyterlab
jupyterlab==2.2.0         # via -r requirements.in
kiwisolver==1.2.0         # via matplotlib
markupsafe==1.1.1         # via jinja2
matplotlib==3.3.0         # via -r requirements.in
mistune==0.8.4            # via nbconvert
nbconvert==5.6.1          # via notebook
nbformat==5.0.7           # via nbconvert, notebook
notebook==6.0.3           # via jupyterlab, jupyterlab-server
numpy==1.19.0             # via -r requirements.in, matplotlib, pandas
packaging==20.4           # via bleach
pandas==1.0.5             # via -r requirements.in
pandocfilters==1.4.2      # via nbconvert
parso==0.7.0              # via jedi
pexpect==4.8.0            # via ipython
pickleshare==0.7.5        # via ipython
pillow==7.2.0             # via matplotlib
prometheus-client==0.8.0  # via notebook
prompt-toolkit==3.0.5     # via ipython
ptyprocess==0.6.0         # via pexpect, terminado
pygments==2.6.1           # via ipython, nbconvert
pyparsing==2.4.7          # via matplotlib, packaging
pyrsistent==0.16.0        # via jsonschema
python-dateutil==2.8.1    # via jupyter-client, matplotlib, pandas
pytz==2020.1              # via pandas
pyzmq==19.0.1             # via jupyter-client, notebook
requests==2.24.0          # via jupyterlab-server
send2trash==1.5.0         # via notebook
six==1.15.0               # via bleach, cycler, jsonschema, packaging, pyrsistent, python-dateutil, traitlets
soupsieve==2.0.1          # via beautifulsoup4
terminado==0.8.3          # via notebook
testpath==0.4.4           # via nbconvert
tornado==6.0.4            # via ipykernel, jupyter-client, jupyterlab, notebook, terminado
traitlets==4.3.3          # via ipykernel, ipython, jupyter-client, jupyter-core, nbconvert, nbformat, notebook
urllib3==1.25.9           # via requests
wcwidth==0.2.5            # via prompt-toolkit
webencodings==0.5.1       # via bleach
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
